Move items, works with lib
