+++ prqlc/prqlc/src/utils/mod.rs
@@ -1,8 +1,9 @@
 mod id_gen;
 mod toposort;
 
-use std::sync::OnceLock;
+use std::{io::stderr, sync::OnceLock};
 
+use anstream::adapter::strip_str;
 pub use id_gen::{IdGenerator, NameGenerator};
 use itertools::Itertools;
 use regex::Regex;
@@ -89,6 +90,26 @@ pub(crate) fn valid_ident() -> &'static Regex {
         // ^ ('*' | [ascii_lower '_$'] [ascii_lower ascii_digit '_$']* ) $
         Regex::new(r"^((\*)|(^[a-z_\$][a-z0-9_\$]*))$").unwrap()
     })
+}
+
+fn should_use_color() -> bool {
+    match anstream::AutoStream::choice(&stderr()) {
+        anstream::ColorChoice::Auto => true,
+        anstream::ColorChoice::Always => true,
+        anstream::ColorChoice::AlwaysAnsi => true,
+        anstream::ColorChoice::Never => false,
+    }
+}
+
+/// Strip colors, for external libraries which don't yet strip themselves, and
+/// for insta snapshot tests. This will respond to environment variables such as
+/// `CLI_COLOR`.
+pub(crate) fn maybe_strip_colors(s: &str) -> String {
+    if !should_use_color() {
+        strip_str(s).to_string()
+    } else {
+        s.to_string()
+    }
 }
 
 #[test]
